Moved new Test to existing class
